feat: extract the Handlebars helpers into a separate module
